report path when reading error in _loadImage()
In case of reading error in _loadImage(), report the error with the path of the file being read. I got an error complaining then data is nil when accessing string.len(data), but I don't know why it happens, whether the file doesn't exist at all, or the file is corrupted.